Jetson/requirements.txt
fastapi
uvicorn
pydantic
requests
numpy
opencv-python
insightface
onnx
onnxruntime-gpu
